+++ mtgcc-spoiler-bot/main.py
@@ -17,7 +17,8 @@ member: Member = Member
 
 
 async def schedule_spoiler_search():
-    from modules.output_to_discord import (build_discord_embed_object, stage_embed_object_for_send, send_notification_to_astral)
+    from modules.output_to_discord import send_text_only_output, stage_embed_object_for_send
+    from modules.create_embed_objects import build_discord_embed_object
 
     current_time = datetime.now()
     next_run = current_time + timedelta(seconds=10)
@@ -31,7 +32,7 @@ async def schedule_spoiler_search():
         await write_card_data_to_file( set_name_data, card_data_object )
         embed_object = await build_discord_embed_object( set_name_data, card_data_object )
         await stage_embed_object_for_send( client, embed_object )
-        await send_notification_to_astral( client, set_name_data, len(embed_object) )
+        await send_text_only_output( client, set_name_data, len(embed_object) )
 
 
 @client.event
